Resolve MIGRATE-ACTION, improve dependencies, plugins and modern CMake use
* Improve header includes / forward declarations to reduce transitive dependencies.

* Split plugin registration source from implementation.

* New INTERFACE targets:

  * larexamples::ShowerCalibrationGaloreTests

  * larexamples::AtomicNumber (provider)

  * larexamples::ShowerCalibrationGaloreService (service interface)

  * larexamples::ShowerCalibrationGalore (provider)

  * larexamples::ShowerCalibrationGaloreProviderTestHelpers

  * larexamples::CheatTrack

* New plugin builders:

  * larexamples::ShowerCalibrationGaloreService

* Use most appropriate plugin builders.
