+++ CMakeLists.txt
@@ -6,12 +6,12 @@
 #  buildtool [-I /install/path] -bt[i] -jN
 # ======================================================================
 
-cmake_minimum_required(VERSION 3.19 FATAL_ERROR)
+cmake_minimum_required(VERSION 3.20 FATAL_ERROR)
 
+find_package(cetmodules REQUIRED)
 project(larexamples VERSION 09.03.15 LANGUAGES CXX)
 
 # cetbuildtools contains our cmake modules
-find_package(cetbuildtools REQUIRED)
 
 include(CetCMakeEnv)
 cet_cmake_env()
@@ -22,42 +22,42 @@ cet_set_compiler_flags(DIAGS CAUTIOUS
   EXTRA_FLAGS -pedantic -Wno-unused-local-typedefs
 )
 
-cet_report_compiler_flags()
+cet_report_compiler_flags(REPORT_THRESHOLD VERBOSE)
 
-find_package(cetlib_except REQUIRED PUBLIC)
-find_package(cetlib REQUIRED PUBLIC)
-find_package(fhiclcpp REQUIRED PUBLIC)
-find_package(messagefacility REQUIRED PUBLIC)
-find_package(canvas_root_io REQUIRED)
-find_package(art REQUIRED PUBLIC)
-find_package(art_root_io REQUIRED PUBLIC)
-find_package(nusimdata REQUIRED PUBLIC)
-find_package(Boost REQUIRED)
-find_package(ROOT COMPONENTS Core  RIO Tree Hist MathCore Physics REQUIRED PUBLIC)
-# these are minimum required versions, not the actual product versions
-find_ups_product( larcoreobj )
-find_ups_product( larcorealg )
-find_ups_product( larcore )
-find_ups_product( lardataobj )
-find_ups_product( lardataalg )
-find_ups_product( lardata )
-find_ups_product( larevt )
-find_ups_product( larsim )
-#find_ups_product( nutools )
+find_package(larcore REQUIRED EXPORT)
+find_package(larcorealg REQUIRED EXPORT)
+find_package(larcoreobj REQUIRED EXPORT)
+find_package(lardata REQUIRED EXPORT)
+find_package(lardataobj REQUIRED EXPORT)
+find_package(larsim REQUIRED EXPORT)
+
+find_package(nusimdata REQUIRED EXPORT)
+
+find_package(art_root_io REQUIRED EXPORT)
+find_package(art REQUIRED EXPORT)
+find_package(messagefacility REQUIRED EXPORT)
+find_package(canvas_root_io REQUIRED EXPORT)
+find_package(canvas REQUIRED EXPORT)
+find_package(fhiclcpp REQUIRED EXPORT)
+find_package(cetlib REQUIRED)
+find_package(cetlib_except REQUIRED)
+
+find_package(ROOT COMPONENTS Core EG Hist MathMore Physics RIO Tree REQUIRED EXPORT)
 
 # macros for dictionary and simple_plugin
-include(ArtDictionary)
+include(BuildDictionary)
 include(ArtMake)
 include(BuildPlugins)
+
+cet_cmake_module_directories(Modules BINARY)
 
 # source
 add_subdirectory(larexamples)
 add_subdirectory(test)
 
 # ups - table and config files
-add_subdirectory(ups)
 
 # packaging utility
-include(UseCPack)
 
 
+cet_cmake_config()
