+++ larexamples/Algorithms/TotallyCheatTracks/CheatTrackData/CMakeLists.txt
@@ -1,5 +1,15 @@
+cet_make_library(LIBRARY_NAME CheatTrack INTERFACE
+  SOURCE CheatTrack.h
+  LIBRARIES INTERFACE
+  lardataobj::RecoBase
+  ROOT::EG
+)
 
-art_dictionary(DICTIONARY_LIBRARIES PRIVATE nusimdata::SimulationBase)
+build_dictionary(DICTIONARY_LIBRARIES PRIVATE
+  larexamples::CheatTrack
+  nusimdata::SimulationBase
+  canvas::canvas
+)
 
 install_source()
 install_headers()
