+++ larexamples/CMakeLists.txt
@@ -1,7 +1,7 @@
-add_subdirectory(AnalysisExample)
-add_subdirectory(Services)
 add_subdirectory(Algorithms)
-add_subdirectory(RecoProxyUsageExample)
+add_subdirectory(AnalysisExample)
 add_subdirectory(DebuggingExamples)
+add_subdirectory(RecoProxyUsageExample)
+add_subdirectory(Services)
 
 install_source()
