+++ larexamples/DebuggingExamples/CatchException/CMakeLists.txt
@@ -1,4 +1,15 @@
-art_make( MODULE_LIBRARIES messagefacility::MF_MessageLogger)
+cet_build_plugin(Disturbance art::EDProducer
+  LIBRARIES PRIVATE
+  canvas::canvas
+  fhiclcpp::types
+)
+
+cet_build_plugin(Exploder art::EDAnalyzer
+  LIBRARIES PRIVATE
+  canvas::canvas
+  messagefacility::MF_MessageLogger
+  fhiclcpp::types
+)
 
 install_source()
 install_fhicl()
