--- larexamples/Services/AtomicNumber/Providers/AtomicNumber.cxx
@@ -1,14 +1,0 @@
-/**
- * @file    AtomicNumber.cxx
- * @brief   Provider returning atomic number of the active material in the TPC
- * @author  Gianluca Petrillo (petrillo@fnal.gov)
- * @date    April 13, 2016
- * @see     AtomicNumber.h
- * @ingroup AtomicNumber
- *
- * Implementation file.
- */
-
-// our own header
-#include "AtomicNumber.h"
-
