+++ larexamples/Services/AtomicNumber/Providers/CMakeLists.txt
@@ -1,6 +1,8 @@
-art_make(
-        LIB_LIBRARIES fhiclcpp::fhiclcpp
-        )
+cet_make_library(LIBRARY_NAME AtomicNumber INTERFACE
+  SOURCE AtomicNumber.h
+  LIBRARIES INTERFACE
+  fhiclcpp::types
+  )
 
 install_fhicl()
 install_headers()
