+++ larexamples/Services/AtomicNumber/Services/CMakeLists.txt
@@ -1,4 +1,9 @@
-art_make()
+cet_build_plugin(AtomicNumberService art::service
+  IMPL_SOURCE
+  LIBRARIES INTERFACE
+  larexamples::AtomicNumber
+  larcore::ServiceProviderWrappers
+)
 
 install_fhicl()
 install_headers()
