+++ test/Services/ShowerCalibrationGalore/Providers/CMakeLists.txt
@@ -1,36 +1,45 @@
+cet_make_library(LIBRARY_NAME ShowerCalibrationGaloreTests INTERFACE
+  NO_INSTALL
+  SOURCE ShowerCalibrationGaloreTests.h
+  LIBRARIES INTERFACE
+  larexamples::ShowerCalibrationGalore
+  lardataobj::RecoBase
+)
+
+cet_make_library(LIBRARY_NAME CreateTestShowerCalibrationFromPID
+  NO_INSTALL
+  SOURCE 
+  CreateTestShowerCalibrationFromPID.cxx
+  LIBRARIES PRIVATE
+  ROOT::Core
+  ROOT::RIO
+  ROOT::Hist
+  ROOT::MathCore
+)
 
 cet_test(CreateTestShowerCalibrationFromPID_test
-  SOURCES CreateTestShowerCalibrationFromPID.cxx CreateTestShowerCalibrationFromPID_test.cc
-  LIBRARIES
-    ROOT::Core
-    ROOT::Hist
-  )
+  SOURCE
+  CreateTestShowerCalibrationFromPID_test.cc
+  LIBRARIES PRIVATE
+  larexamples::CreateTestShowerCalibrationFromPID
+)
 
 cet_test(ShowerCalibrationGaloreFromPID_test
-  SOURCES CreateTestShowerCalibrationFromPID.cxx ShowerCalibrationGaloreFromPID_test.cc
-  LIBRARIES
-    larexamples_Services_ShowerCalibrationGalore_Providers
-    larcorealg_TestUtils
-    messagefacility::MF_MessageLogger
-    
-    fhiclcpp
-    cetlib::cetlib cetlib_except
-    ROOT::Core
-    ROOT::Hist
-  )
+  LIBRARIES PRIVATE
+  larexamples::CreateTestShowerCalibrationFromPID
+  larexamples::ShowerCalibrationGaloreTests
+  larexamples::ShowerCalibrationGaloreProviderTestHelpers
+  larcorealg::unit_test_base
+  messagefacility::MF_MessageLogger
+)
 
 cet_test(ShowerCalibrationGaloreScale_test
-  LIBRARIES
-    larexamples_Services_ShowerCalibrationGalore_Providers
-    larcorealg_TestUtils
-    messagefacility::MF_MessageLogger
-    
-    fhiclcpp
-    cetlib::cetlib cetlib_except
-    ROOT::Core
-    ROOT::Hist
-  )
-
+  LIBRARIES PRIVATE
+  larexamples_Services_ShowerCalibrationGalore_Providers
+  larexamples::ShowerCalibrationGaloreProviderTestHelpers
+  larcorealg::unit_test_base
+  messagefacility::MF_MessageLogger
+)
 
 # install all sources, plus CMakeLists.txt and all configuration files
 file(GLOB TESTFHICLFILES
