--- ups/CMakeLists.txt
@@ -1,6 +1,0 @@
-
-# create package configuration and version files
-
-process_ups_files()
-
-cet_cmake_config()
